move naming and grouping and registrying ESP code into Registrable module
