+++ lib/event_sourcery.rb
@@ -21,8 +21,9 @@ require 'event_sourcery/event_processing/error_handlers/constant_retry'
 require 'event_sourcery/event_processing/error_handlers/exponential_backoff_retry'
 require 'event_sourcery/event_processing/esp_process'
 require 'event_sourcery/event_processing/esp_runner'
-require 'event_sourcery/event_processing/event_stream_processor'
 require 'event_sourcery/event_processing/event_stream_processor_registry'
+require 'event_sourcery/event_processing/registrable'
+require 'event_sourcery/event_processing/event_stream_processor'
 require 'event_sourcery/config'
 require 'event_sourcery/event_body_serializer'
 require 'event_sourcery/aggregate_root'
